Some cleanup
Read from a config.json file for the bot name, description, and token. Also never ever commit your bot's token to GitHub. Please change it now if you haven't done so already.

Reading from a config file in your code, and then adding that config file's name to a .gitignore file is better. I've only added the basis on which to do this by editing the source to read from that config.json file for values

Also #hackathon2018 lol.